backend : filter acces token and basic service work correctly

+++ src/main/java/com/cinema/minute/Service/VideoDkikaService.java
@@ -1,7 +1,9 @@
 package com.cinema.minute.Service;
 
+import com.cinema.minute.Data.Entity.UploadFile;
 import com.cinema.minute.Data.Entity.Videodkika;
 import com.cinema.minute.Data.Repository.VideoDkikaRepo;
+import com.cinema.minute.Service.UploadFile.FileInDB;
 import com.cinema.minute.Service.UploadFile.StorageService;
 import com.cinema.minute.ui.Model.Request.VideoDkikaRequest;
 import com.cinema.minute.ui.Model.Response.videoDkikaResposne;
@@ -10,7 +12,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -19,6 +25,9 @@ import java.util.stream.Collectors;
 public class VideoDkikaService {
     @Autowired
     VideoDkikaRepo videoDkikaRepo;
+
+    @Autowired
+    private FileInDB fileInDB;
 
     @Autowired
     ModelMapper modelMapper;
@@ -31,8 +40,8 @@ public class VideoDkikaService {
         List<videoDkikaResposne> videoDkikaResposneList  = new LinkedList<>();
         for (int i = 0; i < videodkikaList.size()  ; i++) {
             videoDkikaResposne video = modelMapper.map(videodkikaList.get(i),videoDkikaResposne.class);
-            Resource file= storageService.load(videodkikaList.get(i).getVideo().getName());
-            video.setFileInfo(file);
+            if(videodkikaList.get(i).getVideo() != null)
+            video.setIdVideo(videodkikaList.get(i).getVideo().getId());
             videoDkikaResposneList.add(video);
         }
         return videoDkikaResposneList;
@@ -40,23 +49,30 @@ public class VideoDkikaService {
   // @PreAuthorize
     public videoDkikaResposne get(Integer id ) {
         Videodkika v= videoDkikaRepo.findById(id).orElseThrow(()-> new RuntimeException("this video does not exist"));
-        return modelMapper.map(v,videoDkikaResposne.class);
+        videoDkikaResposne videoDkikaResposne = modelMapper.map(v,videoDkikaResposne.class);
+       videoDkikaResposne.setIdVideo(v.getVideo().getId());
+       return videoDkikaResposne;
     }
 
 
     public void addVideo(VideoDkikaRequest Videodkika) {
         Videodkika tt =modelMapper.map(Videodkika,Videodkika.class);
-
+        // get video Uploader using fileInBD service
+        UploadFile fileVideo =fileInDB.getFileById(Videodkika.getVideo());
+         tt.setVideo(fileVideo);
         videoDkikaRepo.saveAndFlush(tt);
     }
 
 
     public void removeVideo(Integer id ){
 
-        // get file id
-        Integer idFile = videoDkikaRepo.findById(id).get().getVideo().getId();
-        storageService.deleteById(idFile);
-        videoDkikaRepo.deleteById(id);
+        Videodkika v  = videoDkikaRepo.findById(id)
+                .orElseThrow(()-> new RuntimeException("video id of video description does not exist"));
+                if(v != null && v.getVideo() == null)
+                    throw new NullPointerException("this video does not exist in data base ");
+                videoDkikaRepo.deleteById(id);
+                storageService.deleteById(v.getVideo().getId());
+
     }
 
     public void updateVideo(VideoDkikaRequest videoDkikaRequest , Integer id){
@@ -64,5 +80,31 @@ public class VideoDkikaService {
         Videodkika vv = modelMapper.map(videoDkikaRequest,Videodkika.class);
         vv.setId(id);
         videoDkikaRepo.saveAndFlush(vv);
+    }
+
+    public Path getPath(Integer id){
+        String p = videoDkikaRepo.findById(id).get().getVideo().getUrlFile();
+       return Paths.get(p);
+    }
+
+
+    public File getResource(Integer id) {
+       Videodkika v =  videoDkikaRepo.findById(id).orElseThrow(()-> new RuntimeException("this file does not exist in data base"));
+       if(v.getVideo() == null)
+           throw new NullPointerException("this video does not exist");
+
+       String url = v.getVideo().getUrlFile();
+       File file = new File(url);
+
+       if(!file.exists())
+           throw new RuntimeException("this file does not exist in folder ");
+       return file;
+    }
+
+    public Integer uploadVideo(MultipartFile file) {
+        UploadFile video =storageService.save(file);
+        if(video == null)
+            throw new NullPointerException("error was eccured in aved video in database or folder");
+        return video.getId();
     }
 }
